Implement a way to automatically repair the order on a delete event
When you delete a model, it will create a gap between the order values
of a model. This merge requests implements a new configuration option
that allows the user to automatically repair this gap by resetting the
order column on all models when a model is deleted.

This is not the most optimal way of doing it. But it is the most
reliable (Even when you've manually set conflicting orders on the
model yourself). In most cases it should work fine though. That's
because orderable models will be, in most cases, in low quantities.

To prevent breaking changes, and improve default performance, I've
disabled the new configuration option by default. So for normal
installations this update should not have any effect.