Update order when a model is deleted
If a model is deleted it leaves a gap in the sort order which might not be desirable in some situations. This change moves a model to the end while deleting it so re-orders all the others.

This PR implements a fix for that but is incomplete in that:
- it doesn't have any tests yet
- it doesn't deal with soft deleted models being restored

If it is of use then I will update it more fully.